requirements.txt
evernote==1.25.3
beautifulsoup4==4.6.0
python-decouple==3.5
